requirements_v4.txt
# PicLocate V4 Production Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Google Drive & Auth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Computer Vision & ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLO
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.22.0
scikit-learn>=1.3.0
imagehash>=4.3.0  # Perceptual hashing

# CLIP for embeddings (install separately if needed)
ftfy>=6.1.0
regex>=2023.0.0
tqdm>=4.65.0

# Supabase & Database
supabase==2.0.3
postgrest==0.13.0
realtime==1.0.0
gotrue==1.3.0
storage3==0.7.4

# OpenAI for embeddings & translations
openai==1.3.5

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# HTTP & Utilities
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.1

# Optional: Advanced Features
# transformers==4.35.2  # For Grounding DINO / other transformers models
# segment-anything==1.0  # For SAM2 segmentation (upgrade later)

